# doc/swagger/simple-bank.swagger.yaml
swagger: "2.0"
info:
  title: rpc_create_user.proto
  version: version not set
tags:
  - name: SimpleBank
consumes:
  - application/json
produces:
  - application/json
paths:
  /v1/login:
    post:
      operationId: SimpleBank_LoginUser
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/pbLoginUserResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/pbLoginUserRequest'
      tags:
        - SimpleBank
  /v1/users:
    post:
      operationId: SimpleBank_CreateUser
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/pbCreateUserResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/pbCreateUserRequest'
      tags:
        - SimpleBank
    patch:
      operationId: SimpleBank_UpdateUser
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/pbUpdateUserResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/pbUpdateUserRequest'
      tags:
        - SimpleBank
definitions:
  pbCreateUserRequest:
    type: object
    properties:
      username:
        type: string
      fullName:
        type: string
      email:
        type: string
      password:
        type: string
  pbCreateUserResponse:
    type: object
    properties:
      user:
        $ref: '#/definitions/pbUser'
  pbLoginUserRequest:
    type: object
    properties:
      username:
        type: string
      password:
        type: string
  pbLoginUserResponse:
    type: object
    properties:
      sessionId:
        type: string
      accessToken:
        type: string
      refreshToken:
        type: string
      accessTokenExpiresAt:
        type: string
        format: date-time
      refreshTokenExpiresAt:
        type: string
        format: date-time
      user:
        $ref: '#/definitions/pbUser'
  pbUpdateUserRequest:
    type: object
    properties:
      username:
        type: string
      fullName:
        type: string
      email:
        type: string
      password:
        type: string
  pbUpdateUserResponse:
    type: object
    properties:
      user:
        $ref: '#/definitions/pbUser'
  pbUser:
    type: object
    properties:
      username:
        type: string
      fullName:
        type: string
      email:
        type: string
      passwordChangedAt:
        type: string
        format: date-time
      createdAt:
        type: string
        format: date-time
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
    additionalProperties: {}
  rpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          $ref: '#/definitions/protobufAny'